+++ pkg/convert/apkbuild/apkbuild_test.go
@@ -1,4 +1,4 @@
-package convert
+package apkbuild
 
 import (
 	"bytes"
@@ -10,6 +10,8 @@ import (
 	"strings"
 	"testing"
 	"time"
+
+	"chainguard.dev/melange/pkg/util"
 
 	rlhttp "chainguard.dev/melange/pkg/http"
 	"chainguard.dev/melange/pkg/manifest"
@@ -36,7 +38,7 @@ func TestGetApkDependencies(t *testing.T) {
 	server := httptest.NewServer(http.HandlerFunc(func(rw http.ResponseWriter, req *http.Request) {
 
 		// assert requests dependency is in the list of test files
-		assert.True(t, Contains(filenames, req.URL.String()), "requests file does not match any test files")
+		assert.True(t, util.Contains(filenames, req.URL.String()), "requests file does not match any test files")
 
 		// send response to be tested
 		data, err := os.ReadFile(filepath.Join("testdata", "deps", "/"+req.URL.String()))
