+++ pkg/convert/apkbuild/testdata/cheese-7.8.9.tar.bz2
