+++ pkg/convert/apkbuild/testdata/multi_source/APKBUILD
