+++ pkg/convert/apkbuild/testdata/scanner_error.yaml
