+++ pkg/convert/python/testdata/meta/botocore.json
