+++ pkg/convert/python/testdata/meta/six.json
